Move Geometry.h to utils
Geometry.h is used in many places that have nothing to do with guilib.
Also, the classes in it are generic and not limited to GUI tasks in
their purpose. This commit moves the file to the general utils
directory so this intention is made clear and there are no
dependencies on guilib sprayed over the code going forward.
