+++ xbmc/cores/RetroPlayer/guicontrols/GUIGameControl.cpp
@@ -26,11 +26,11 @@
 #include "cores/RetroPlayer/rendering/RenderSettings.h"
 #include "cores/RetroPlayer/rendering/RenderVideoSettings.h"
 #include "games/GameServices.h"
-#include "guilib/Geometry.h"
 #include "guilib/GraphicContext.h"
 #include "guilib/TransformMatrix.h"
 #include "settings/GameSettings.h"
 #include "settings/MediaSettings.h"
+#include "utils/Geometry.h"
 #include "utils/StringUtils.h"
 #include "Application.h"
 #include "ApplicationPlayer.h"
