+++ xbmc/cores/RetroPlayer/rendering/IRenderManager.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "guilib/Geometry.h"
+#include "utils/Geometry.h"
 
 struct RESOLUTION_INFO;
