+++ xbmc/cores/RetroPlayer/rendering/RenderGeometry.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "guilib/Geometry.h"
+#include "utils/Geometry.h"
 
 namespace KODI
 {
