+++ xbmc/cores/VideoPlayer/DVDInputStreams/DVDInputStreamNavigator.cpp
@@ -25,8 +25,8 @@
 #include "settings/Settings.h"
 #include "LangInfo.h"
 #include "ServiceBroker.h"
+#include "utils/Geometry.h"
 #include "utils/log.h"
-#include "guilib/Geometry.h"
 #include "utils/URIUtils.h"
 #include "utils/StringUtils.h"
 #include "guilib/LocalizeStrings.h"
