+++ xbmc/cores/VideoPlayer/VideoRenderers/RenderManager.h
@@ -24,7 +24,7 @@
 
 #include "cores/VideoPlayer/VideoRenderers/BaseRenderer.h"
 #include "cores/VideoPlayer/VideoRenderers/OverlayRenderer.h"
-#include "guilib/Geometry.h"
+#include "utils/Geometry.h"
 #include "guilib/Resolution.h"
 #include "threads/CriticalSection.h"
 #include "cores/VideoSettings.h"
