+++ xbmc/guilib/VisibleEffect.h
@@ -30,7 +30,7 @@ class Tweener;
 class CGUIListItem;
 
 #include "TransformMatrix.h"  // needed for the TransformMatrix member
-#include "Geometry.h"         // for CPoint, CRect
+#include "utils/Geometry.h"         // for CPoint, CRect
 #include <memory>
 #include "interfaces/info/InfoBool.h"
