+++ xbmc/platform/android/activity/XBMCApp.h
@@ -38,13 +38,13 @@
 #include "threads/Event.h"
 #include "interfaces/IAnnouncer.h"
 
-#include "guilib/Geometry.h"
 #include "IActivityHandler.h"
 #include "IInputHandler.h"
 #include "JNIMainActivity.h"
 #include "JNIXBMCAudioManagerOnAudioFocusChangeListener.h"
 #include "JNIXBMCMediaSession.h"
 #include "platform/xbmc.h"
+#include "utils/Geometry.h"
 
 // forward declares
 class CJNIWakeLock;
