+++ xbmc/platform/win32/WIN32Util.h
@@ -24,7 +24,7 @@
 
 #include "URL.h"
 #include "MediaSource.h"
-#include "guilib/Geometry.h"
+#include "utils/Geometry.h"
 
 #define BONJOUR_EVENT             ( WM_USER + 0x100 )	// Message sent to the Window when a Bonjour event occurs.
 #define BONJOUR_BROWSER_EVENT     ( WM_USER + 0x110 )
