+++ xbmc/utils/Geometry.h
