+++ xbmc/windowing/wayland/WindowDecorator.h
@@ -27,11 +27,11 @@
 #include <wayland-cursor.hpp>
 
 #include "Connection.h"
-#include "guilib/Geometry.h"
 #include "Registry.h"
 #include "ShellSurface.h"
 #include "threads/CriticalSection.h"
 #include "Util.h"
+#include "utils/Geometry.h"
 #include "utils/posix/SharedMemory.h"
 #include "WindowDecorationHandler.h"
